backend/requirements.txt
langchain 
langchain-community
supabase
openai 
fastapi 
pydantic 
python-dotenv
python-multipart